+++ EssentialFeed/EssentialFeedTests/Feed Cache/Helpers/FeedStoreSpy.swift
@@ -1,0 +1,54 @@
+//
+//  FeedStoreSpy.swift
+//  EssentialFeedTests
+//
+//  Created by Aaron Huánuco on 13/09/2020.
+//  Copyright © 2020 Aaron Huánuco. All rights reserved.
+//
+
+import Foundation
+import EssentialFeed
+
+class FeedStoreSpy: FeedStore {
+    typealias DeletionCompletion = (Error?) -> Void
+    typealias InsertionCompletion = (Error?) -> Void
+
+    var deleteCacheFeedCallCount = 0
+
+    enum ReceivedMessages: Equatable {
+        case deleteCacheFeed
+        case insert([LocalFeedImage], Date)
+    }
+
+    private(set) var receivedMessages = [ReceivedMessages]()
+    private var deletionCompletions = [DeletionCompletion]()
+    private var insertionCompletions = [InsertionCompletion]()
+
+    // MARK: - Deletion
+    func deleteCachedFeed(completion: @escaping DeletionCompletion) {
+        deletionCompletions.append(completion)
+        receivedMessages.append(.deleteCacheFeed)
+    }
+
+    func completeDeletion(with error: Error, at index: Int = 0) {
+        deletionCompletions[index](error)
+    }
+
+    func completeDeletionSuccessfully(at index: Int = 0) {
+        deletionCompletions[index](nil)
+    }
+
+    // MARK: - Insertion
+    func insert(_ feed: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
+        insertionCompletions.append(completion)
+        receivedMessages.append(.insert(feed, timestamp))
+    }
+
+    func completeInsertion(with error: Error, at index: Int = 0) {
+        insertionCompletions[index](error)
+    }
+
+    func completeInsertionSuccessfully(at index: Int = 0) {
+         insertionCompletions[index](nil)
+    }
+}
